Begin fleshing out of White Paper, in Markdown
Flesh out intro to try and clarify need for this work. Provide
sections to give an overview of each part of the FNAL stack for
illustration of the issues.

Make changes in markdown (cause I'm having trouble grokking org...)
and then flush back to org using pandoc. This doesn't give fully
idempotent conversions, so provide org conversion as a separate
file:

- wp.md main markdown changes
- wp-frommd.org, conversion via pandoc of above